requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
scikit-learn==1.7.2
joblib==1.4.2
python-Levenshtein==0.27.1
pymongo==4.9.1
motor==3.6.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.0
email-validator==2.2.0
